+++ c_api/fastdeploy_capi/vision/detection/ppdet/model.cc
@@ -14,7 +14,7 @@
 
 #include "fastdeploy_capi/vision/detection/ppdet/model.h"
 
-#include "fastdeploy_capi/types_internal.h"
+#include "fastdeploy_capi/internal/types_internal.h"
 #include "fastdeploy_capi/vision/visualize.h"
 
 #ifdef __cplusplus
