+++ c_api/fastdeploy_capi/vision/visualize.cc
@@ -15,7 +15,7 @@
 #include "fastdeploy_capi/vision/visualize.h"
 
 #include "fastdeploy/vision/visualize/visualize.h"
-#include "fastdeploy_capi/types_internal.h"
+#include "fastdeploy_capi/internal/types_internal.h"
 
 #ifdef __cplusplus
 extern "C" {
