+++ examples/vision/detection/paddledetection/c/infer_ppyoloe.c
@@ -36,7 +36,7 @@ void CpuInfer(const char* model_dir, const char* image_file) {
   FD_C_RuntimeOptionWrapperUseCpu(option);
 
   FD_C_PPYOLOEWrapper* model = FD_C_CreatePPYOLOEWrapper(
-      model_file, params_file, config_file, option, PADDLE);
+      model_file, params_file, config_file, option, FD_C_ModelFormat_PADDLE);
 
   if (!FD_C_PPYOLOEWrapperInitialized(model)) {
     printf("Failed to initialize.\n");
@@ -84,7 +84,7 @@ void GpuInfer(const char* model_dir, const char* image_file) {
   FD_C_RuntimeOptionWrapperUseGpu(option, 0);
 
   FD_C_PPYOLOEWrapper* model = FD_C_CreatePPYOLOEWrapper(
-      model_file, params_file, config_file, option, PADDLE);
+      model_file, params_file, config_file, option, FD_C_ModelFormat_PADDLE);
 
   if (!FD_C_PPYOLOEWrapperInitialized(model)) {
     printf("Failed to initialize.\n");
