+++ examples/vision/segmentation/paddleseg/cpu-gpu/c/infer.c
@@ -36,7 +36,7 @@ void CpuInfer(const char* model_dir, const char* image_file) {
   FD_C_RuntimeOptionWrapperUseCpu(option);
 
   FD_C_PaddleSegModelWrapper* model = FD_C_CreatePaddleSegModelWrapper(
-      model_file, params_file, config_file, option, PADDLE);
+      model_file, params_file, config_file, option, FD_C_ModelFormat_PADDLE);
 
   if (!FD_C_PaddleSegModelWrapperInitialized(model)) {
     printf("Failed to initialize.\n");
@@ -88,7 +88,7 @@ void GpuInfer(const char* model_dir, const char* image_file) {
   FD_C_RuntimeOptionWrapperUseGpu(option, 0);
 
   FD_C_PaddleSegModelWrapper* model = FD_C_CreatePaddleSegModelWrapper(
-      model_file, params_file, config_file, option, PADDLE);
+      model_file, params_file, config_file, option, FD_C_ModelFormat_PADDLE);
 
   if (!FD_C_PaddleSegModelWrapperInitialized(model)) {
     printf("Failed to initialize.\n");
